fix: defer identity exit journal work
