+++ packages/kit-bg/src/dbs/simple/entity/SimpleDbEntityPrime.ts
@@ -134,6 +134,30 @@ export type IRemoteOneKeyIdLogoutPresentationClaimResult =
       status: 'handled' | 'unavailable';
     };
 
+export type IClaimRemoteOneKeyIdLogoutPresentationParams = {
+  operationId: string;
+  messageId: string;
+  claimId: string;
+  expiresAt: number;
+  now: number;
+};
+
+export type ICompleteRemoteOneKeyIdLogoutPresentationParams = {
+  operationId: string;
+  messageId: string;
+  claimId: string;
+  presentationHandledAt: number;
+  tombstoneExpiresAt: number;
+};
+
+export type IUpdateRemoteOneKeyIdLogoutJournalDeliveryParams = {
+  operationId: string;
+  messageId: string;
+  acknowledgedAt?: number;
+  presentationHandledAt?: number;
+  tombstoneExpiresAt?: number;
+};
+
 export type IKeylessOAuthSessionPersistenceJournal = {
   operationId: string;
   status: 'prepared';
@@ -431,48 +455,9 @@ export class SimpleDbEntityPrime extends SimpleDbEntityBase<ISimpleDBPrime> {
   async setIdentityExitJournalEntry(
     entry: IIdentityExitJournalEntry,
   ): Promise<void> {
-    await this.setRawData((rawData) => {
-      const existing =
-        rawData?.identityExitOperationJournal?.[entry.operationId];
-      const existingDelivery = existing?.remoteDeviceLogout;
-      const nextDelivery = entry.remoteDeviceLogout;
-      const mergedEntry =
-        existingDelivery &&
-        nextDelivery &&
-        existingDelivery.messageId === nextDelivery.messageId
-          ? {
-              ...entry,
-              remoteDeviceLogout: {
-                ...nextDelivery,
-                acknowledgedAt:
-                  existingDelivery.acknowledgedAt ??
-                  nextDelivery.acknowledgedAt,
-                presentationHandledAt:
-                  existingDelivery.presentationHandledAt ??
-                  nextDelivery.presentationHandledAt,
-                presentationHandledClaimId:
-                  existingDelivery.presentationHandledClaimId ??
-                  nextDelivery.presentationHandledClaimId,
-                presentationClaim:
-                  existingDelivery.presentationHandledAt ||
-                  nextDelivery.presentationHandledAt
-                    ? undefined
-                    : (existingDelivery.presentationClaim ??
-                      nextDelivery.presentationClaim),
-                tombstoneExpiresAt:
-                  existingDelivery.tombstoneExpiresAt ??
-                  nextDelivery.tombstoneExpiresAt,
-              },
-            }
-          : entry;
-      return {
-        ...rawData,
-        identityExitOperationJournal: {
-          ...rawData?.identityExitOperationJournal,
-          [entry.operationId]: mergedEntry,
-        },
-      };
-    });
+    const { setIdentityExitJournalEntry } =
+      await import('./identityExitJournal');
+    await setIdentityExitJournalEntry(this, entry);
   }
 
   @backgroundMethod()
@@ -509,200 +494,28 @@ export class SimpleDbEntityPrime extends SimpleDbEntityBase<ISimpleDBPrime> {
   }
 
   @backgroundMethod()
-  async updateRemoteOneKeyIdLogoutJournalDelivery({
-    operationId,
-    messageId,
-    acknowledgedAt,
-    presentationHandledAt,
-    tombstoneExpiresAt,
-  }: {
-    operationId: string;
-    messageId: string;
-    acknowledgedAt?: number;
-    presentationHandledAt?: number;
-    tombstoneExpiresAt?: number;
-  }): Promise<IIdentityExitJournalEntry | undefined> {
-    let result: IIdentityExitJournalEntry | undefined;
-    await this.setRawData((rawData) => {
-      const entry = rawData?.identityExitOperationJournal?.[operationId];
-      if (entry?.remoteDeviceLogout?.messageId !== messageId) {
-        return { ...rawData };
-      }
-
-      const remoteDeviceLogout = {
-        ...entry.remoteDeviceLogout,
-      };
-      let changed = false;
-      if (
-        remoteDeviceLogout.acknowledgedAt === undefined &&
-        acknowledgedAt !== undefined
-      ) {
-        remoteDeviceLogout.acknowledgedAt = acknowledgedAt;
-        changed = true;
-      }
-      if (
-        entry.status === 'completed' &&
-        remoteDeviceLogout.presentationHandledAt === undefined &&
-        presentationHandledAt !== undefined
-      ) {
-        remoteDeviceLogout.presentationHandledAt = presentationHandledAt;
-        remoteDeviceLogout.presentationClaim = undefined;
-        changed = true;
-      }
-      if (
-        remoteDeviceLogout.acknowledgedAt !== undefined &&
-        remoteDeviceLogout.presentationHandledAt !== undefined &&
-        remoteDeviceLogout.tombstoneExpiresAt === undefined &&
-        tombstoneExpiresAt !== undefined
-      ) {
-        remoteDeviceLogout.tombstoneExpiresAt = tombstoneExpiresAt;
-        changed = true;
-      }
-
-      if (!changed) {
-        result = entry;
-        return { ...rawData };
-      }
-      result = {
-        ...entry,
-        updatedAt: Math.max(Date.now(), entry.updatedAt + 1),
-        remoteDeviceLogout,
-      };
-      return {
-        ...rawData,
-        identityExitOperationJournal: {
-          ...rawData?.identityExitOperationJournal,
-          [operationId]: result,
-        },
-      };
-    });
-    return result;
+  async updateRemoteOneKeyIdLogoutJournalDelivery(
+    params: IUpdateRemoteOneKeyIdLogoutJournalDeliveryParams,
+  ): Promise<IIdentityExitJournalEntry | undefined> {
+    const { updateRemoteOneKeyIdLogoutJournalDelivery } =
+      await import('./identityExitJournal');
+    return updateRemoteOneKeyIdLogoutJournalDelivery(this, params);
   }
 
-  async tryClaimRemoteOneKeyIdLogoutPresentation({
-    operationId,
-    messageId,
-    claimId,
-    expiresAt,
-    now,
-  }: {
-    operationId: string;
-    messageId: string;
-    claimId: string;
-    expiresAt: number;
-    now: number;
-  }): Promise<IRemoteOneKeyIdLogoutPresentationClaimResult> {
-    let result: IRemoteOneKeyIdLogoutPresentationClaimResult = {
-      status: 'unavailable',
-    };
-    await this.setRawData((rawData) => {
-      const entry = rawData?.identityExitOperationJournal?.[operationId];
-      const delivery = entry?.remoteDeviceLogout;
-      if (
-        entry?.status !== 'completed' ||
-        !entry.completed?.oneKeyIdLoggedOut ||
-        delivery?.messageId !== messageId
-      ) {
-        return { ...rawData };
-      }
-      if (delivery.presentationHandledAt) {
-        result = { status: 'handled' };
-        return { ...rawData };
-      }
-      if (
-        delivery.presentationClaim &&
-        delivery.presentationClaim.expiresAt > now
-      ) {
-        result = {
-          status: 'claimedByOther',
-          retryAfterMs: delivery.presentationClaim.expiresAt - now,
-        };
-        return { ...rawData };
-      }
-
-      const nextEntry: IIdentityExitJournalEntry = {
-        ...entry,
-        updatedAt: Math.max(now, entry.updatedAt + 1),
-        remoteDeviceLogout: {
-          ...delivery,
-          presentationClaim: {
-            claimId,
-            expiresAt,
-          },
-        },
-      };
-      result = {
-        status: 'claimed',
-        claimId,
-        expiresAt,
-      };
-      return {
-        ...rawData,
-        identityExitOperationJournal: {
-          ...rawData?.identityExitOperationJournal,
-          [operationId]: nextEntry,
-        },
-      };
-    });
-    return result;
+  async tryClaimRemoteOneKeyIdLogoutPresentation(
+    params: IClaimRemoteOneKeyIdLogoutPresentationParams,
+  ): Promise<IRemoteOneKeyIdLogoutPresentationClaimResult> {
+    const { tryClaimRemoteOneKeyIdLogoutPresentation } =
+      await import('./identityExitJournal');
+    return tryClaimRemoteOneKeyIdLogoutPresentation(this, params);
   }
 
-  async completeRemoteOneKeyIdLogoutPresentation({
-    operationId,
-    messageId,
-    claimId,
-    presentationHandledAt,
-    tombstoneExpiresAt,
-  }: {
-    operationId: string;
-    messageId: string;
-    claimId: string;
-    presentationHandledAt: number;
-    tombstoneExpiresAt: number;
-  }): Promise<IIdentityExitJournalEntry | undefined> {
-    let result: IIdentityExitJournalEntry | undefined;
-    await this.setRawData((rawData) => {
-      const entry = rawData?.identityExitOperationJournal?.[operationId];
-      const delivery = entry?.remoteDeviceLogout;
-      if (
-        entry?.status !== 'completed' ||
-        !entry.completed?.oneKeyIdLoggedOut ||
-        delivery?.messageId !== messageId
-      ) {
-        return { ...rawData };
-      }
-      if (delivery.presentationHandledAt) {
-        if (delivery.presentationHandledClaimId === claimId) {
-          result = entry;
-        }
-        return { ...rawData };
-      }
-      if (delivery.presentationClaim?.claimId !== claimId) {
-        return { ...rawData };
-      }
-
-      result = {
-        ...entry,
-        updatedAt: Math.max(presentationHandledAt, entry.updatedAt + 1),
-        remoteDeviceLogout: {
-          ...delivery,
-          presentationHandledAt,
-          presentationHandledClaimId: claimId,
-          presentationClaim: undefined,
-          tombstoneExpiresAt: delivery.acknowledgedAt
-            ? (delivery.tombstoneExpiresAt ?? tombstoneExpiresAt)
-            : delivery.tombstoneExpiresAt,
-        },
-      };
-      return {
-        ...rawData,
-        identityExitOperationJournal: {
-          ...rawData?.identityExitOperationJournal,
-          [operationId]: result,
-        },
-      };
-    });
-    return result;
+  async completeRemoteOneKeyIdLogoutPresentation(
+    params: ICompleteRemoteOneKeyIdLogoutPresentationParams,
+  ): Promise<IIdentityExitJournalEntry | undefined> {
+    const { completeRemoteOneKeyIdLogoutPresentation } =
+      await import('./identityExitJournal');
+    return completeRemoteOneKeyIdLogoutPresentation(this, params);
   }
 
   async removeIdentityExitJournalEntry({
